# perf: Switch the STROBE implementation