--- a/LICENSE
+++ b/LICENSE
@@ -1,6 +1,6 @@
 Copyright (c) 2016-2019 isis agora lovecruft. All rights reserved.
 Copyright (c) 2016-2019 Henry de Valence. All rights reserved.
-Copyright (c) 2016, 2019 The Go Authors. All rights reserved.
+Copyright (c) 2014, 2015, 2016, 2019 The Go Authors. All rights reserved.
 Copyright (c) 2017, 2019 George Tankersley. All rights reserved.
 Copyright (c) 2019-2020 Web 3 Foundation. All rights reserved.
 Copyright (c) 2020 Jack Grigg. All rights reserved.