--- a/go.mod
+++ b/go.mod
@@ -3,7 +3,6 @@ module github.com/oasisprotocol/curve25519-voi
 go 1.16
 
 require (
-	github.com/mimoo/StrobeGo v0.0.0-20181016162300-f8f6d4d2b643 // indirect
 	golang.org/x/crypto v0.0.0-20201221181555-eec23a3978ad
 	golang.org/x/sys v0.0.0-20191026070338-33540a1f6037
 )