This PR implements a minimal subset of STROBE to replace the external dependency.  While the existing implementation appears to be correct, it is extremely allocation heavy, especially when used to implement merlin due to the lack of `more` support for operations, and the gratuitous copies.

TODO:
 * [x] Implement STROBE
 * [x] Switch merlin to use the new implementation
 * [x] Do some light optimization (add a fast path to duplex)
 * [x] Add some tests beyond "well merlin still works"

Fixes #61