@Yawning 2021-06-07
nb: Can't be bothered disabling turbo or running multiple iterations, and messages used for the benchmarks are all "short".

```
name                    old time/op    new time/op    delta
VerifyBatchOnly/1-8       45.0µs ± 0%    42.6µs ± 0%   -5.41%
VerifyBatchOnly/2-8       63.2µs ± 0%    59.0µs ± 0%   -6.66%
VerifyBatchOnly/4-8        100µs ± 0%      92µs ± 0%   -8.23%
VerifyBatchOnly/8-8        174µs ± 0%     158µs ± 0%   -9.49%
VerifyBatchOnly/16-8       315µs ± 0%     289µs ± 0%   -8.36%
VerifyBatchOnly/32-8       602µs ± 0%     554µs ± 0%   -7.96%
VerifyBatchOnly/64-8      1.20ms ± 0%    1.10ms ± 0%   -8.20%
VerifyBatchOnly/128-8     2.12ms ± 0%    1.97ms ± 0%   -7.27%
VerifyBatchOnly/256-8     3.95ms ± 0%    3.61ms ± 0%   -8.69%
VerifyBatchOnly/384-8     5.68ms ± 0%    5.15ms ± 0%   -9.36%
VerifyBatchOnly/512-8     7.45ms ± 0%    6.77ms ± 0%   -9.12%
VerifyBatchOnly/768-8     11.3ms ± 0%     9.7ms ± 0%  -13.89%
VerifyBatchOnly/1024-8    14.0ms ± 0%    12.6ms ± 0%   -9.87%
GenerateKeyPair-8         12.6µs ± 0%    12.6µs ± 0%   -0.21%
Signing-8                 21.1µs ± 0%    16.3µs ± 0%  -23.00%
Verification-8            31.4µs ± 0%    28.9µs ± 0%   -8.10%

name                    old alloc/op   new alloc/op   delta
VerifyBatchOnly/1-8       9.08kB ± 0%    6.92kB ± 0%  -23.79%
VerifyBatchOnly/2-8       15.2kB ± 0%    11.8kB ± 0%  -22.30%
VerifyBatchOnly/4-8       28.4kB ± 0%    22.6kB ± 0%  -20.59%
VerifyBatchOnly/8-8       53.2kB ± 0%    42.5kB ± 0%  -20.22%
VerifyBatchOnly/16-8       111kB ± 0%      90kB ± 0%  -18.57%
VerifyBatchOnly/32-8       212kB ± 0%     172kB ± 0%  -18.98%
VerifyBatchOnly/64-8       419kB ± 0%     340kB ± 0%  -18.97%
VerifyBatchOnly/128-8      478kB ± 0%     320kB ± 0%  -33.09%
VerifyBatchOnly/256-8      946kB ± 0%     631kB ± 0%  -33.32%
VerifyBatchOnly/384-8     1.54MB ± 0%    1.07MB ± 0%  -30.63%
VerifyBatchOnly/512-8     1.89MB ± 0%    1.26MB ± 0%  -33.37%
VerifyBatchOnly/768-8     3.07MB ± 0%    2.13MB ± 0%  -30.73%
VerifyBatchOnly/1024-8    3.73MB ± 0%    2.47MB ± 0%  -33.77%
GenerateKeyPair-8           368B ± 0%      368B ± 0%    0.00%
Signing-8                 2.45kB ± 0%    0.80kB ± 0%  -67.32%
Verification-8              664B ± 0%       32B ± 0%  -95.18%

name                    old allocs/op  new allocs/op  delta
VerifyBatchOnly/1-8         59.0 ± 0%      19.0 ± 0%  -67.80%
VerifyBatchOnly/2-8         96.0 ± 0%      26.0 ± 0%  -72.92%
VerifyBatchOnly/4-8          169 ± 0%        39 ± 0%  -76.92%
VerifyBatchOnly/8-8          314 ± 0%        64 ± 0%  -79.62%
VerifyBatchOnly/16-8         603 ± 0%       113 ± 0%  -81.26%
VerifyBatchOnly/32-8       1.18k ± 0%     0.21k ± 0%  -82.20%
VerifyBatchOnly/64-8       2.33k ± 0%     0.40k ± 0%  -82.73%
VerifyBatchOnly/128-8      4.64k ± 0%     0.79k ± 0%  -82.99%
VerifyBatchOnly/256-8      9.25k ± 0%     1.56k ± 0%  -83.15%
VerifyBatchOnly/384-8      13.9k ± 0%      2.3k ± 0%  -83.21%
VerifyBatchOnly/512-8      18.5k ± 0%      3.1k ± 0%  -83.24%
VerifyBatchOnly/768-8      27.7k ± 0%      4.6k ± 0%  -83.27%
VerifyBatchOnly/1024-8     36.9k ± 0%      6.2k ± 0%  -83.28%
GenerateKeyPair-8           6.00 ± 0%      6.00 ± 0%    0.00%
Signing-8                   38.0 ± 0%      12.0 ± 0%  -68.42%
Verification-8              15.0 ± 0%       1.0 ± 0%  -93.33%
```

Note that for large message sizes ("unreasonably large" on amd64) this will be slower, but the answer is "use `sr25519.NewTranscriptHash` instead of trying to sign gigantic messages".  If people happen to care about arm64 performance, they are more than welcome to submit a followup PR, the required improvements are obvious.